requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.5.2
sqlmodel==0.0.21
httpx==0.27.2
langchain==0.3.7
langchain-groq==0.2.1
langchain-chroma==0.1.4
langchain-community==0.3.7
langchain-core==0.3.17
langgraph==0.2.38
langsmith==0.1.136
chromadb==0.5.15
sentence-transformers==3.1.1
slack-sdk==3.33.1
google-api-python-client==2.149.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
ragas==0.2.3
datasets==2.21.0
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-httpx==0.32.0
python-dotenv==1.0.1
